backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
MetaTrader5==5.0.4200
numpy<2
python-dotenv==1.1.0
supabase==2.11.0
